Refactor docker-compose.yml to improve routing and load balancing
The changes to the docker-compose configuration file enhance the routing and load balancing functionalities of Traefik, thus improving handling of incoming traffic. New configuration options for the entrypoint and port mapping have been added. This includes the --entrypoints.web.address=:80 option for establishing a new entrypoint that listens on port 80, along with "80:80" mapping to bind port 80 of the container to the host's port 80. These changes facilitate efficient traffic routing to the host's port 80, thus promoting effective networking and simplifying potential troubleshooting.