Update environment variables in docker-compose
Replaced hard-coded values with environment variables in docker-compose.yml for flexibility and security. Now, `LE_EMAIL` is used instead of specific email for certificate resolver and `CHAT_DOMAIN` is used for routing rule, making it easier to manage configurations for different environments and securing sensitive data.

In addition, this change makes it necessary to update README with instructions on setting up these variables.