Update OpenAI model configurations
Updated the chat bot configurations to a new OpenAI model (GPT-4) and set its default temperature. This is because GPT-4 is more performant and handles more conversational tokens compared to the previous GPT_3_5 models. Also the ModelSelect was disabled due to this change since only one model is being actively used.